Implement source specific crpix/crval pairs.
This is so that the pixel/sky coordinate transformation will still be correct in the presence of distortions when using the local affine transformation.  The local linear transformation is expanded around the inital ra/dec of each source; this simply sets the reference coordinate and the reference pixel to be the center of the source, accounting for distortions.